ADP-4003 -- Refactor some code
