+++ yarn-error.log
@@ -1,0 +1,90 @@
+Arguments: 
+  /Users/darko/.nvm/versions/node/v9.8.0/bin/node /usr/local/Cellar/yarn/1.5.1_1/libexec/bin/yarn.js run test
+
+PATH: 
+  /Users/darko/.nvm/versions/node/v9.8.0/bin:/usr/local/bin:/usr/bin:/bin:/usr/sbin:/sbin:/opt/X11/bin:/Users/darko/.antigen/bundles/chrissicool/zsh-256color:/Users/darko/.antigen/bundles/unixorn/autoupdate-antigen.zshplugin:/Users/darko/.antigen/bundles/robbyrussell/oh-my-zsh/plugins/command-not-found:/Users/darko/.antigen/bundles/robbyrussell/oh-my-zsh/plugins/common-aliases:/Users/darko/.antigen/bundles/djui/alias-tips:/Users/darko/.antigen/bundles/robbyrussell/oh-my-zsh/plugins/git:/Users/darko/.antigen/bundles/robbyrussell/oh-my-zsh/plugins/heroku:/Users/darko/.antigen/bundles/robbyrussell/oh-my-zsh/plugins/extract:/Users/darko/.antigen/bundles/zlsun/solarized-man:/Users/darko/.antigen/bundles/zsh-users/zsh-syntax-highlighting:/Users/darko/.antigen/bundles/zsh-users/zsh-history-substring-search
+
+Yarn version: 
+  1.5.1
+
+Node version: 
+  9.8.0
+
+Platform: 
+  darwin x64
+
+npm manifest: 
+  {
+    "name": "react-combo-select",
+    "version": "1.2.8",
+    "keywords": "react, select, singleselect, multiselect, search, react-component",
+    "description": "React select component that can be used as a single select, multiselect, search",
+    "main": "./lib/Components/ComboSelect.js",
+    "scripts": {
+      "start": "./node_modules/.bin/webpack-dev-server --config webpack.config.js  --hot --progress --colors --port 8080 --open",
+      "test": "mocha -r mock-local-storage --compilers js:babel-core/register --require ./test/test.js \"test/**/*@(.js|.jsx)\"",
+      "prepublish": "rm -rf lib && mkdir lib && ./node_modules/babel-cli/bin/babel.js src --out-dir lib"
+    },
+    "repository": {
+      "type": "git",
+      "url": "git+https://github.com/gogoair/react-combo-select.git"
+    },
+    "author": "Marko Kostovski <marko.kostovski@gmail.com>",
+    "license": "MIT",
+    "bugs": {
+      "url": "https://github.com/gogoair/react-combo-select/issues"
+    },
+    "homepage": "https://github.com/gogoair/react-combo-select#readme",
+    "devDependencies": {
+      "babel-cli": "^6.6.5",
+      "babel-core": "^6.7.4",
+      "babel-eslint": "^6.0.2",
+      "babel-loader": "^6.2.4",
+      "babel-preset-env": "^1.6.0",
+      "babel-preset-react": "^6.5.0",
+      "babel-preset-stage-0": "^6.24.1",
+      "chai": "^3.5.0",
+      "chai-immutable": "^1.5.4",
+      "enzyme": "^3.1.0",
+      "enzyme-adapter-react-16": "^1.0.1",
+      "eslint": "^4.18.1",
+      "eslint-plugin-react": "^4.2.3",
+      "immutable": "^3.8.2",
+      "jsdom": "^8.3.0",
+      "mocha": "^4.0.1",
+      "mock-local-storage": "^1.0.2",
+      "react-hot-loader": "^3.0.0-beta.7",
+      "react-test-renderer": "^16.0.0",
+      "sinon": "^1.17.4",
+      "webpack": "^1.12.14",
+      "webpack-dev-server": "^1.14.1"
+    },
+    "dependencies": {
+      "prop-types": "^15.6.0"
+    },
+    "peerDependencies": {
+      "react": "*",
+      "react-dom": "*"
+    }
+  }
+
+yarn manifest: 
+  No manifest
+
+Lockfile: 
+  No lockfile
+
+Trace: 
+  Error: Command failed.
+  Exit code: 1
+  Command: sh
+  Arguments: -c mocha -r mock-local-storage --compilers js:babel-core/register --require ./test/test.js "test/**/*@(.js|.jsx)"
+  Directory: /Users/darko/Projects/react-combo-select
+  Output:
+  
+      at ProcessTermError.MessageError (/usr/local/Cellar/yarn/1.5.1_1/libexec/lib/cli.js:186:110)
+      at new ProcessTermError (/usr/local/Cellar/yarn/1.5.1_1/libexec/lib/cli.js:226:113)
+      at ChildProcess.<anonymous> (/usr/local/Cellar/yarn/1.5.1_1/libexec/lib/cli.js:30281:17)
+      at ChildProcess.emit (events.js:180:13)
+      at maybeClose (internal/child_process.js:936:16)
+      at Process.ChildProcess._handle.onexit (internal/child_process.js:220:5)
